+++ qmake/generators/win32/msvc_vcproj.cpp
@@ -1117,7 +1117,7 @@ void VcprojGenerator::initResourceFiles()
                     if(FILE *proc = QT_POPEN(dep_cmd.latin1(), "r")) {
         	        QString indeps;
                         while(!feof(proc)) {
-                            int read_in = fread(buff, 1, 255, proc);
+                            int read_in = (int)fread(buff, 1, 255, proc);
                             if(!read_in)
                                 break;
                             indeps += QByteArray(buff, read_in);
